Update github actions pipeline
