+++ .env.template
@@ -4,12 +4,12 @@
 JAVA_HOME=/opt/jdk25/temurin
 M2_HOME=
 MAVEN_USER_HOME=$M2_HOME
-MAVEN_HOME=$M2_HOME
+MAVEN_HOME=$HOME/icecream/.local-volumes/maven-3.9.12
 ##########################################
 ### LOCAL MACHINE
 ##########################################
 LOCAL_VOLUME=
-LOCAL_STORAGE=
+LOCAL_SYSTEM=
 ##########################################
 ### POSTGRESQL
 ##########################################
