+++ .github/workflows/actions.yaml
@@ -1,0 +1,121 @@
+name: Build & Push All Images
+
+on:
+  push:
+    branches: ["master"]
+
+env:
+  REGISTRY: ghcr.io
+  IMAGE_PREFIX: ${{ github.repository_owner }}/icecream
+
+jobs:
+
+  #region Backend services
+  backend:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        service: [consul, chronos, conflux, echo, memoir, codex]
+    permissions:
+      contents: read
+      packages: write
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Delete old package versions
+        uses: actions/delete-package-versions@v5
+        with:
+          package-name: icecream-${{ matrix.service }}
+          package-type: container
+          min-versions-to-keep: 0
+          delete-only-untagged-versions: false
+        continue-on-error: true
+      - name: Login to GHCR
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+      - name: Build & Push Backend Image
+        uses: docker/build-push-action@v5
+        with:
+          context: ./icecream-system
+          file: ./icecream-system/Dockerfile
+          push: true
+          build-args: |
+            MODULE=${{ matrix.service }}
+          tags: |
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-${{ matrix.service }}:latest
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-${{ matrix.service }}:${{ github.sha }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+  #endregion
+
+  #region UI Storefront
+  frontend:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Delete old package versions
+        uses: actions/delete-package-versions@v5
+        with:
+          package-name: icecream-storefront
+          package-type: container
+          min-versions-to-keep: 0
+          delete-only-untagged-versions: false
+        continue-on-error: true
+      - name: Login to GHCR
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+      - name: Build & Push Frontend Image
+        uses: docker/build-push-action@v5
+        with:
+          context: ./icecream-ui/storefront
+          file: ./icecream-ui/storefront/Dockerfile
+          push: true
+          tags: |
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-storefront:latest
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-storefront:${{ github.sha }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+  #endregion
+
+  #region Infra
+  infra-identity:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Delete old package versions
+        uses: actions/delete-package-versions@v5
+        with:
+          package-name: icecream-infra-identity
+          package-type: container
+          min-versions-to-keep: 0
+          delete-only-untagged-versions: false
+        continue-on-error: true
+      - name: Login to GHCR
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+      - name: Build & Push Identity Image
+        uses: docker/build-push-action@v6
+        with:
+          context: ./icecream-system/identity
+          push: true
+          tags: |
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-infra-identity:latest
+            ${{ env.REGISTRY }}/${{ env.IMAGE_PREFIX }}-infra-identity:${{ github.sha }}
+          # This makes subsequent builds lightning fast
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+  #endregion
