+++ icecream-system/.env.template
@@ -1,4 +1,4 @@
-JAVA_TOOL_OPTIONS="-javaagent:${LOCAL_STORAGE}/java-packages/opentelemetry-javaagent.jar -XX:NativeMemoryTracking=summary"
+JAVA_TOOL_OPTIONS="-javaagent:${LOCAL_SYSTEM}/java-packages/opentelemetry-javaagent.jar -XX:NativeMemoryTracking=summary"
 OTEL_EXPORTER_OTLP_ENDPOINT=http://${OTEL_COLLECTOR_HOST}:${OTEL_COLLECTOR_PORT}
 OTEL_EXPORTER_OTLP_PROTOCOL=http/protobuf
 OTEL_LOGS_EXPORTER=otlp
