+++ icecream-system/dev-composes/compose.yml
@@ -154,24 +154,6 @@ services:
       retries: 5
       start_period: 30s
 
-#  kibana:
-#    image: kibana:8.19.6
-#    container_name: icecream-kibana-optional
-#    ports:
-#      - "${KIBANA_PORT}:5601"
-#    environment:
-#      - ELASTICSEARCH_HOSTS=http://elasticsearch:9200
-#    depends_on:
-#      - elasticsearch
-#    networks:
-#      - icecream-network
-#    healthcheck:
-#      test: [ "CMD-SHELL", "curl -f http://localhost:5601/api/status || exit 1" ]
-#      interval: 10s
-#      timeout: 10s
-#      retries: 5
-#      start_period: 30s
-
   zookeeper:
     image: confluentinc/cp-zookeeper:7.8.4
     container_name: icecream-zookeeper
