+++ icecream-system/identity/Dockerfile
@@ -1,0 +1,17 @@
+FROM quay.io/keycloak/keycloak:26.2.5 AS builder
+
+ENV KC_HEALTH_ENABLED=true
+ENV KC_METRICS_ENABLED=true
+ENV KC_DB=postgres
+
+WORKDIR /opt/keycloak
+COPY ./themes /opt/keycloak/themes/
+RUN /opt/keycloak/bin/kc.sh build
+
+FROM quay.io/keycloak/keycloak:26.2.5
+COPY --from=builder /opt/keycloak/ /opt/keycloak/
+
+COPY ./data/import/realm-export.json /opt/keycloak/data/import/
+
+ENTRYPOINT ["/opt/keycloak/bin/kc.sh"]
+CMD ["start", "--optimized", "--import-realm"]
