+++ icecream-ui/storefront/.env.template
@@ -1,15 +1,10 @@
-NEXT_PUBLIC_CONSUL_URL=${CONSUL_URL}
-NEXT_PUBLIC_ECHO_URL=${ECHO_URL}
+CONSUL_URL=${CONSUL_URL}
+ECHO_URL=${ECHO_URL}
+STORAGE_URL=${STORAGE_URL}
 AUTH_SECRET=${STOREFRONT_AUTH_SECRET}
 AUTH_KEYCLOAK_ID=${IDENTITY_STOREFRONT_CLIENT_ID}
 AUTH_KEYCLOAK_SECRET=${IDENTITY_STOREFRONT_CLIENT_SECRET}
 AUTH_KEYCLOAK_ISSUER=${IDENTITY_ISSUER}
-NEXT_PUBLIC_STORAGE_URL=${STORAGE_URL}
 STORAGE_ACCESS_KEY=${MINIO_ACCESS_KEY}
 STORAGE_SECRET_KEY=${MINIO_SECRET_KEY}
 STORAGE_IMAGE_BUCKET=images
-OTEL_EXPORTER_OTLP_ENDPOINT=http://${OTEL_COLLECTOR_HOST}:${OTEL_COLLECTOR_PORT}
-OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://${OTEL_COLLECTOR_HOST}:${OTEL_COLLECTOR_PORT}/v1/traces
-OTEL_EXPORTER_OTLP_TRACES_PROTOCOL=http/protobuf
-OTEL_LOG_LEVEL=debug
-NEXT_OTEL_VERBOSE=1
