+++ icecream-ui/storefront/src/repositories/consul/orders.ts
@@ -2,7 +2,7 @@ import { API_PATHS } from '@/lib/constants'
 import { api } from '@/lib/rest-client'
 import { generateUrl } from '@/lib/utils'
 import { CreateOrderRequest, CreateOrderResponse, Session } from '@/models'
-import { CONSUL_URL } from '@/settings'
+import { CONSUL_URL } from '@/lib/env'
 
 export const requestCreateOrder = async (
   session: Session,
