+++ icecream-ui/storefront/src/repositories/consul/products.ts
@@ -2,7 +2,7 @@ import { API_PATHS } from '@/lib/constants'
 import { api } from '@/lib/rest-client'
 import { generateUrl } from '@/lib/utils'
 import { ProductExtended, RequestAllParams, RequestAllResult } from '@/models'
-import { CONSUL_URL } from '@/settings'
+import { CONSUL_URL } from '@/lib/env'
 
 const _requestAllProducts = async (
   { pagination, sorting, filters }: RequestAllParams,
