--- icecream-ui/storefront/src/settings.ts
@@ -1,9 +1,0 @@
-export const CONSUL_URL = process.env.NEXT_PUBLIC_CONSUL_URL as string
-export const ECHO_URL = process.env.NEXT_PUBLIC_ECHO_URL as string
-export const STORAGE_URL = process.env.NEXT_PUBLIC_STORAGE_URL as string
-export const AUTH_KEYCLOAK_ID = process.env.AUTH_KEYCLOAK_ID as string
-export const AUTH_KEYCLOAK_SECRET = process.env.AUTH_KEYCLOAK_SECRET as string
-export const AUTH_KEYCLOAK_ISSUER = process.env.AUTH_KEYCLOAK_ISSUER as string
-export const STORAGE_ACCESS_KEY = process.env.STORAGE_ACCESS_KEY as string
-export const STORAGE_SECRET_KEY = process.env.STORAGE_SECRET_KEY as string
-export const STORAGE_IMAGE_BUCKET = process.env.STORAGE_IMAGE_BUCKET as string
