+++ readme-assets/architecture.png
