+++ readme-assets/observability-architecture.png
